requirements.txt
# Core dependencies for Advanced RAG System

# Web framework
streamlit>=1.28.0

# LangChain ecosystem
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1

# Document processing
PyPDF2>=3.0.0
pdfplumber>=0.9.0
python-docx>=0.8.11

# Vector stores and embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# BM25 for keyword search
rank-bm25>=0.2.2

# Groq API client
groq>=0.4.1

# OpenAI (optional, for embeddings)
openai>=1.3.0

# Text processing and utilities
numpy>=1.24.0
pandas>=2.0.0
tiktoken>=0.5.0

# Logging and monitoring
python-dotenv>=1.0.0

# Optional: Unstructured for better document parsing
unstructured[pdf]>=0.10.0

# Development and testing dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
